+++ orx-jvm/orx-boofcv/src/main/kotlin/PointConversion.kt
@@ -7,5 +7,5 @@ import org.openrndr.math.Vector2
 
 fun Point2D_I32.toVector2() = Vector2(x.toDouble(), y.toDouble())
 fun Point2D_F32.toVector2() = Vector2(x.toDouble(), y.toDouble())
-fun Point2D_F64.toVector2() = Vector2(x.toDouble(), y.toDouble())
+fun Point2D_F64.toVector2() = Vector2(x, y)
 fun List<Point2D_I32>.toVector2s() = this.map { it.toVector2()  }
